Editor: Move the InterfaceSkeleton to the editor package
