--- packages/edit-post/src/components/header/style.scss
@@ -1,14 +1,0 @@
-/**
- *  Show icon label overrides.
- */
-.show-icon-labels .editor-header {
-	.edit-post-fullscreen-mode-close.has-icon {
-		width: $header-height;
-		svg {
-			display: block;
-		}
-		&::after {
-			content: none;
-		}
-	}
-}
