+++ packages/edit-post/src/components/more-menu/manage-patterns-menu-item.js
