+++ packages/editor/src/private-apis.js
@@ -11,6 +11,8 @@ import { lock } from './lock-unlock';
 import { EntitiesSavedStatesExtensible } from './components/entities-saved-states';
 import useAutoSwitchEditorSidebars from './components/provider/use-auto-switch-editor-sidebars';
 import useBlockEditorSettings from './components/provider/use-block-editor-settings';
+import BackButton from './components/header/back-button';
+import EditorCanvas from './components/editor-canvas';
 import Header from './components/header';
 import CreateTemplatePartModal from './components/create-template-part-modal';
 import InserterSidebar from './components/inserter-sidebar';
@@ -43,9 +45,11 @@ const { store: interfaceStore, ...remainingInterfaceApis } = interfaceApis;
 export const privateApis = {};
 lock( privateApis, {
 	CreateTemplatePartModal,
+	BackButton,
 	ExperimentalEditorProvider,
 	EntitiesSavedStatesExtensible,
 	GlobalStylesProvider,
+	EditorCanvas,
 	Header,
 	InserterSidebar,
 	ListViewSidebar,
